+++ src/hangar/remotes.py
@@ -10,6 +10,7 @@ from typing import List, NamedTuple, Optional, Sequence
 import grpc
 import lmdb
 from tqdm import tqdm
+import numpy as np
 
 from .backends import backend_decoder
 from .constants import LMDB_SETTINGS
@@ -364,7 +365,6 @@ class Remotes(object):
                             while notEmpty:
                                 notEmpty = curs.delete()
                     unpack_commit_ref(self._env.refenv, tmpDB, commit)
-                    print(f'Client processing commit {commit}')
                     recQuery = queries.RecordQuery(tmpDB)
 
                     # handle column_names option
@@ -374,14 +374,13 @@ class Remotes(object):
                     else:
                         cmt_columns = [col for col in column_names if col in cmt_column_names]
                     for col in cmt_columns:
-                        print(f'Client Querying Hashs for Column: {col}')
                         cmtData_hashs = recQuery.column_data_hashes(col)
                         allHashs.update(cmtData_hashs)
             finally:
                 tmpDB.close()
 
-        hashTxn = TxnRegister().begin_reader_txn(self._env.hashenv)
         try:
+            hashTxn = TxnRegister().begin_reader_txn(self._env.hashenv)
             m_schema_hash_map = defaultdict(list)
             for hashVal in allHashs:
                 hashKey = hash_data_db_key_from_raw_key(hashVal.digest)
@@ -406,7 +405,10 @@ class Remotes(object):
                     # max_num_bytes option
                     if isinstance(max_num_bytes, int):
                         for idx, r_kv in enumerate(ret):
-                            total_nbytes_seen += r_kv[1].nbytes
+                            try:
+                                total_nbytes_seen += r_kv[1].nbytes
+                            except AttributeError:
+                                total_nbytes_seen += len(r_kv[1])
                             if total_nbytes_seen >= max_num_bytes:
                                 ret = ret[0:idx]
                                 stop = True